Title: Use Panda to play the background video instead of CEF
Playing this video in CEF appears to be very expensive (could be due to
how the offscreen buffer/texture is handled) and it results in about a
20ms frametime hit on an Intel HD 630 and about a 14ms hit on an AMD
Vega M GL. Unfortunately, this means pulling in ffmpeg (a 20mb
dependency) just to play a title background. At the moment the disk size
hit is preferable to the frame time hit.